Bookmarks and disc sort
Added bookmarks
Added Jukebox by default server option
Added option to clear playlist and completion of playlist playback
Added option to sort by disc and track
Moved remove all and save playlist to a new row of buttons
Moved download status text to ActionBar subtitle
Added opaque background for visualizer

+++ src/com/thejoshwa/ultrasonic/androidapp/service/CachedMusicService.java
@@ -385,7 +385,7 @@ public class CachedMusicService implements MusicService {
 	}
 
 	@Override
-	public void createBookmark(String id, Long position, Context context, ProgressListener progressListener) throws Exception {
+	public void createBookmark(String id, int position, Context context, ProgressListener progressListener) throws Exception {
 		musicService.createBookmark(id, position, context, progressListener);
 	}
 }
